refactor(Names): 플레이어의 수 검증 역할 위임
